feat: optimize image seeding logic and add option to omit images
This commit optimizes the image seeding logic in the `seedPlacesFromWorkbook` method of the `SeedsService` class. It introduces an option to omit images when seeding places from a workbook. The new `omit-images` query parameter allows users to control whether images should be created or recreated during the seeding process. By default, images are created or recreated, but when the `omit-images` parameter is set to `true`, no images will be processed.

These changes improve the flexibility and efficiency of the seeding process, allowing users to choose whether to include images based on their specific requirements.