refactor: update lodging table columns
This commit adds new columns to the lodging table in the database. The 'phone' column is dropped, and the following columns are added: 'room_types' (an array of text), 'phones' (an array of text), 'facebook' (text), 'instagram' (text), and 'whataapps' (an array of text). The migration file '1726497162426-add-new-columns-to-lodging-table.ts' is created to handle these changes.